+++ src/Microsoft.AspNetCore.JsonPatch/Adapters/ObjectAdapter.cs
@@ -34,66 +34,6 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
         /// </summary>
         public Action<JsonPatchError> LogErrorAction { get; }
 
-        /// <summary>
-        /// The "add" operation performs one of the following functions,
-        /// depending upon what the target location references:
-        ///
-        /// o  If the target location specifies an array index, a new value is
-        ///    inserted into the array at the specified index.
-        ///
-        /// o  If the target location specifies an object member that does not
-        ///    already exist, a new member is added to the object.
-        ///
-        /// o  If the target location specifies an object member that does exist,
-        ///    that member's value is replaced.
-        ///
-        /// The operation object MUST contain a "value" member whose content
-        /// specifies the value to be added.
-        ///
-        /// For example:
-        ///
-        /// { "op": "add", "path": "/a/b/c", "value": [ "foo", "bar" ] }
-        ///
-        /// When the operation is applied, the target location MUST reference one
-        /// of:
-        ///
-        /// o  The root of the target document - whereupon the specified value
-        ///    becomes the entire content of the target document.
-        ///
-        /// o  A member to add to an existing object - whereupon the supplied
-        ///    value is added to that object at the indicated location.  If the
-        ///    member already exists, it is replaced by the specified value.
-        ///
-        /// o  An element to add to an existing array - whereupon the supplied
-        ///    value is added to the array at the indicated location.  Any
-        ///    elements at or above the specified index are shifted one position
-        ///    to the right.  The specified index MUST NOT be greater than the
-        ///    number of elements in the array.  If the "-" character is used to
-        ///    index the end of the array (see [RFC6901]), this has the effect of
-        ///    appending the value to the array.
-        ///
-        /// Because this operation is designed to add to existing objects and
-        /// arrays, its target location will often not exist.  Although the
-        /// pointer's error handling algorithm will thus be invoked, this
-        /// specification defines the error handling behavior for "add" pointers
-        /// to ignore that error and add the value as specified.
-        ///
-        /// However, the object itself or an array containing it does need to
-        /// exist, and it remains an error for that not to be the case.  For
-        /// example, an "add" with a target location of "/a/b" starting with this
-        /// document:
-        ///
-        /// { "a": { "foo": 1 } }
-        ///
-        /// is not an error, because "a" exists, and "b" will be added to its
-        /// value.  It is an error in this document:
-        ///
-        /// { "q": { "bar": 2 } }
-        ///
-        /// because "a" does not exist.
-        /// </summary>
-        /// <param name="operation">The add operation.</param>
-        /// <param name="objectToApplyTo">Object to apply the operation to.</param>
         public void Add(Operation operation, object objectToApplyTo)
         {
             if (operation == null)
@@ -138,7 +78,7 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             var visitor = new ObjectVisitor(parsedPath, ContractResolver);
 
             var target = objectToApplyTo;
-            if (!visitor.TryVisit(ref target, adapterWithTest: out var adapter, errorMessage: out var errorMessage))
+            if (!visitor.TryVisit(ref target, out var adapter, out var errorMessage))
             {
                 var error = CreatePathNotFoundError(objectToApplyTo, path, operation, errorMessage);
                 ErrorReporter(error);
@@ -153,29 +93,6 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             }
         }
 
-        /// <summary>
-        /// The "move" operation removes the value at a specified location and
-        /// adds it to the target location.
-        ///
-        /// The operation object MUST contain a "from" member, which is a string
-        /// containing a JSON Pointer value that references the location in the
-        /// target document to move the value from.
-        ///
-        /// The "from" location MUST exist for the operation to be successful.
-        ///
-        /// For example:
-        ///
-        /// { "op": "move", "from": "/a/b/c", "path": "/a/b/d" }
-        ///
-        /// This operation is functionally identical to a "remove" operation on
-        /// the "from" location, followed immediately by an "add" operation at
-        /// the target location with the value that was just removed.
-        ///
-        /// The "from" location MUST NOT be a proper prefix of the "path"
-        /// location; i.e., a location cannot be moved into one of its children.
-        /// </summary>
-        /// <param name="operation">The move operation.</param>
-        /// <param name="objectToApplyTo">Object to apply the operation to.</param>
         public void Move(Operation operation, object objectToApplyTo)
         {
             if (operation == null)
@@ -202,20 +119,6 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             }
         }
 
-        /// <summary>
-        /// The "remove" operation removes the value at the target location.
-        ///
-        /// The target location MUST exist for the operation to be successful.
-        ///
-        /// For example:
-        ///
-        /// { "op": "remove", "path": "/a/b/c" }
-        ///
-        /// If removing an element from an array, any elements above the
-        /// specified index are shifted one position to the left.
-        /// </summary>
-        /// <param name="operation">The remove operation.</param>
-        /// <param name="objectToApplyTo">Object to apply the operation to.</param>
         public void Remove(Operation operation, object objectToApplyTo)
         {
             if (operation == null)
@@ -244,7 +147,7 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             var visitor = new ObjectVisitor(parsedPath, ContractResolver);
 
             var target = objectToApplyTo;
-            if (!visitor.TryVisit(ref target, adapterWithTest: out var adapter, errorMessage: out var errorMessage))
+            if (!visitor.TryVisit(ref target, out var adapter, out var errorMessage))
             {
                 var error = CreatePathNotFoundError(objectToApplyTo, path, operationToReport, errorMessage);
                 ErrorReporter(error);
@@ -259,26 +162,6 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             }
         }
 
-        /// <summary>
-        /// The "replace" operation replaces the value at the target location
-        /// with a new value.  The operation object MUST contain a "value" member
-        /// whose content specifies the replacement value.
-        ///
-        /// The target location MUST exist for the operation to be successful.
-        ///
-        /// For example:
-        ///
-        /// { "op": "replace", "path": "/a/b/c", "value": 42 }
-        ///
-        /// This operation is functionally identical to a "remove" operation for
-        /// a value, followed immediately by an "add" operation at the same
-        /// location with the replacement value.
-        ///
-        /// Note: even though it's the same functionally, we do not call remove + add
-        /// for performance reasons (multiple checks of same requirements).
-        /// </summary>
-        /// <param name="operation">The replace operation.</param>
-        /// <param name="objectToApplyTo">Object to apply the operation to.</param>
         public void Replace(Operation operation, object objectToApplyTo)
         {
             if (operation == null)
@@ -295,7 +178,7 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             var visitor = new ObjectVisitor(parsedPath, ContractResolver);
 
             var target = objectToApplyTo;
-            if (!visitor.TryVisit(ref target, adapterWithTest: out var adapter, errorMessage: out var errorMessage))
+            if (!visitor.TryVisit(ref target, out var adapter, out var errorMessage))
             {
                 var error = CreatePathNotFoundError(objectToApplyTo, operation.path, operation, errorMessage);
                 ErrorReporter(error);
@@ -310,28 +193,6 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             }
         }
 
-        /// <summary>
-        ///  The "copy" operation copies the value at a specified location to the
-        ///  target location.
-        ///
-        ///  The operation object MUST contain a "from" member, which is a string
-        ///  containing a JSON Pointer value that references the location in the
-        ///  target document to copy the value from.
-        ///
-        ///  The "from" location MUST exist for the operation to be successful.
-        ///
-        ///  For example:
-        ///
-        ///  { "op": "copy", "from": "/a/b/c", "path": "/a/b/e" }
-        ///
-        ///  This operation is functionally identical to an "add" operation at the
-        ///  target location using the value specified in the "from" member.
-        ///
-        ///  Note: even though it's the same functionally, we do not call add with
-        ///  the value specified in from for performance reasons (multiple checks of same requirements).
-        /// </summary>
-        /// <param name="operation">The copy operation.</param>
-        /// <param name="objectToApplyTo">Object to apply the operation to.</param>
         public void Copy(Operation operation, object objectToApplyTo)
         {
             if (operation == null)
@@ -365,24 +226,6 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             }
         }
 
-        /// <summary>
-        /// The "test" operation tests that a value at the target location
-        /// is equal to a specified value.
-        /// 
-        /// The operation object MUST contain a "value" member that conveys the 
-        /// value to be compared to the target location's value.
-        /// 
-        /// The target location MUST be equal to the "value" value for the 
-        /// operation to be considered successful.
-        /// 
-        /// Also, note that ordering of the serialization of object members 
-        /// is not significant.
-        /// 
-        /// For example:
-        /// { "op": "test", "path": "/a/b/c", "value": "foo" }
-        /// </summary>
-        /// <param name="operation">The test operation.</param>
-        /// <param name="objectToApplyTo">Object to apply the operation to.</param>
         public void Test(Operation operation, object objectToApplyTo)
         {
             if (operation == null)
@@ -399,7 +242,7 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             var visitor = new ObjectVisitor(parsedPath, ContractResolver);
 
             var target = objectToApplyTo;
-            if (!visitor.TryVisit(ref target, adapterWithTest: out var adapter, errorMessage: out var errorMessage))
+            if (!visitor.TryVisit(ref target, out var adapter, out var errorMessage))
             {
                 var error = CreatePathNotFoundError(objectToApplyTo, operation.path, operation, errorMessage);
                 ErrorReporter(error);
@@ -441,7 +284,7 @@ namespace Microsoft.AspNetCore.JsonPatch.Adapters
             var visitor = new ObjectVisitor(parsedPath, ContractResolver);
 
             var target = objectToGetValueFrom;
-            if (!visitor.TryVisit(ref target, adapterWithTest: out var adapter, errorMessage: out var errorMessage))
+            if (!visitor.TryVisit(ref target, out var adapter, out var errorMessage))
             {
                 var error = CreatePathNotFoundError(objectToGetValueFrom, fromLocation, operation, errorMessage);
                 ErrorReporter(error);
