+++ src/Microsoft.AspNetCore.JsonPatch/Operations/Operation.cs
@@ -58,44 +58,15 @@ namespace Microsoft.AspNetCore.JsonPatch.Operations
                     adapter.Copy(this, objectToApplyTo);
                     break;
                 case OperationType.Test:
-                    throw new NotSupportedException(Resources.TestOperationNotSupported);
-                default:
-                    break;
-            }
-        }
-
-        public void Apply(object objectToApplyTo, IObjectAdapterWithTest adapter)
-        {
-            if (objectToApplyTo == null)
-            {
-                throw new ArgumentNullException(nameof(objectToApplyTo));
-            }
-
-            if (adapter == null)
-            {
-                throw new ArgumentNullException(nameof(adapter));
-            }
-
-            switch (OperationType)
-            {
-                case OperationType.Add:
-                    adapter.Add(this, objectToApplyTo);
-                    break;
-                case OperationType.Remove:
-                    adapter.Remove(this, objectToApplyTo);
-                    break;
-                case OperationType.Replace:
-                    adapter.Replace(this, objectToApplyTo);
-                    break;
-                case OperationType.Move:
-                    adapter.Move(this, objectToApplyTo);
-                    break;
-                case OperationType.Copy:
-                    adapter.Copy(this, objectToApplyTo);
-                    break;
-                case OperationType.Test:
-                    adapter.Test(this, objectToApplyTo);
-                    break;
+                    if (adapter is IObjectAdapterWithTest adapterWithTest)
+                    {
+                        adapterWithTest.Test(this, objectToApplyTo);
+                        break;
+                    }
+                    else
+                    {
+                        throw new NotSupportedException(Resources.TestOperationNotSupported);
+                    }
                 default:
                     break;
             }
